Attempt to fix iPad bug
